basic基本通过 进行到case5，rest有问题

+++ src/main.cpp
@@ -22,10 +22,20 @@ bool isValidInt(std::string& str){
     }
     return true;
 }
+std::vector<std::string> get_keyword(std::string str){
+    std::vector<std::string> substrings;
+    std::istringstream iss1(str);
+    std::string sub;
+    while(std::getline(iss1, sub,'|')){
+        substrings.push_back(sub);
+    }
+    if(substrings.empty()){substrings.push_back(str);}
+    return substrings;
+}
 int main(){
 
-//   freopen("/home/hqs123/bookstore/src/test/basic/testcase3.in","r",stdin);
-//   freopen("0.out","w",stdout);
+   freopen("/home/hqs123/bookstore/src/test/basic/testcase5.in","r",stdin);
+   freopen("0.out","w",stdout);
 
     Booksystem booksystem("isbn_data","isbn_index","isbn_tail",
                         "name_data","name_index","name_tail",
@@ -35,7 +45,13 @@ int main(){
     Finance finance("finance");
 
     std::string input_string;
+    int num_count=1;
     while(std::getline(std::cin,input_string)){
+        std::cout<<num_count;
+        if(num_count==36){
+            int aaa=19;
+        }
+        num_count++;
         bool result = std::all_of(input_string.begin(), input_string.end(), [](char c) { return c == ' '; });
         if(result) continue;//先处理一下空串
         std::istringstream iss(input_string);
@@ -51,31 +67,36 @@ int main(){
         if(command=="su"){
             if(v.size()==2) {usersystem.su(v[0],v[1]);}
             else if(v.size()==1) {usersystem.su(v[0],"");}
-            else {std::cout<<"Invalid\n";continue;}
+            else {std::cout<<"Invalid"<<std::endl;continue;}
         }
         else if(command=="logout"){
+            if(usersystem.login_now.empty()){std::cout<<"Invalid"<<std::endl;continue;}
             if(v.empty()) {usersystem.logout();}
-            else {std::cout<<"Invalid\n";continue;}
+            else {std::cout<<"Invalid"<<std::endl;continue;}
         }
         else if(command=="register"){
             if(v.size()==3) {usersystem.register_(v[0],v[1],v[2]);}
-            else {std::cout<<"Invalid\n";continue;}
+            else {std::cout<<"Invalid"<<std::endl;continue;}
         }
         else if(command=="passwd"){
+            //错误：要求登录状态下才能进行这个操作
+            if(usersystem.login_now.empty()){std::cout<<"Invalid"<<std::endl;continue;}
             if(v.size()==3) {usersystem.modify(v[0],v[2],v[1]);}
             else if(v.size()==2) {usersystem.modify(v[0],v[1],"");}
-            else {std::cout<<"Invalid\n";continue;}
+            else {std::cout<<"Invalid"<<std::endl;continue;}
         }
         else if(command=="useradd"){
+            if(usersystem.login_now.empty()){std::cout<<"Invalid"<<std::endl;continue;}
             if(v.size()==4 && (v[2]=="0"||v[2]=="1"||v[2]=="3"||v[2]=="7")){
                 int tmp=std::stoi(v[2]);
                 usersystem.useradd(v[0],v[1],v[3],tmp);
             }
-            else {std::cout<<"Invalid\n";continue;}
+            else {std::cout<<"Invalid"<<std::endl;continue;}
         }
         else if(command=="delete"){
+            if(usersystem.login_now.empty()){std::cout<<"Invalid"<<std::endl;continue;}
             if(v.size()==1){usersystem.delete_(v[0]);}
-            else {std::cout<<"Invalid\n";continue;}
+            else {std::cout<<"Invalid"<<std::endl;continue;}
         }
         else if(command=="show"){//两种情况
             if(v.size()==2 && v[0]=="finance" && isValidInt(v[1])==true){
@@ -100,7 +121,7 @@ int main(){
                 std::sort(books.begin(), books.end());
                 for(int i=0;i<books.size();++i){
                     std::cout<<books[i].isbn<<'\t'<<books[i].BookName<<'\t'<<books[i].Author
-                    <<'\t'<<books[i].Keyword<<'\t'<<books[i].price<<'\t'<<books[i].rest<<'\n';
+                    <<'\t'<<books[i].Keyword<<'\t'<<books[i].price<<'\t'<<books[i].rest<<std::endl;
                 }
             }
             else if(v.size()==1 && v[0]!="finance"){
@@ -108,21 +129,51 @@ int main(){
                 // 不确定：[Keyword] 中出现多个关键词则操作失败？
 
                 int pos=v[0].find("=");
-                if(pos==std::string::npos) {std::cout<<"Invalid\n";continue;}
+                if(pos==std::string::npos) {std::cout<<"Invalid"<<std::endl;continue;}
                 if(pos==v[0].size()-1) {std::cout<<"Invalid\n";break;}
                 std::string s1=v[0].substr(0,pos);
-                std::string s2=v[0].substr(pos+1);
-                if(s1=="-ISBN"){booksystem.show_isbn(s2);}
-                else if(s1=="-name"){booksystem.show_name(s2);}
-                else if(s1=="-author"){booksystem.show_author(s2);}
-                else if(s1=="-keyword"){booksystem.show_keyword(s2);}
-                else {std::cout<<"Invalid\n";continue;}   
+                //不确定：这里引号没有怎么处理？
+                if(s1=="-ISBN"){
+                    std::string s2=v[0].substr(pos+1);
+                    booksystem.show_isbn(s2);
+                }
+                else if(s1=="-name"){
+                    std::string s2;
+                    int startPos = v[0].find("\"");       // 查找第一个引号的位置
+                    int endPos = v[0].find_last_of("\""); // 查找最后一个引号的位置
+                    if (startPos != std::string::npos && endPos != std::string::npos && startPos < endPos) {
+                        s2=v[0].substr(startPos + 1, endPos - startPos - 1); // 提取引号中的内容
+                    }
+                    else{std::cout<<"Invalid"<<std::endl;continue;}
+                    booksystem.show_name(s2);
+                }
+                else if(s1=="-author"){
+                    std::string s2;
+                    int startPos = v[0].find("\"");       // 查找第一个引号的位置
+                    int endPos = v[0].find_last_of("\""); // 查找最后一个引号的位置
+                    if (startPos != std::string::npos && endPos != std::string::npos && startPos < endPos) {
+                        s2=v[0].substr(startPos + 1, endPos - startPos - 1); // 提取引号中的内容
+                    }
+                    else{std::cout<<"Invalid"<<std::endl;continue;}
+                    booksystem.show_author(s2);
+                }
+                else if(s1=="-keyword"){
+                    std::string s2;
+                    int startPos = v[0].find("\"");       // 查找第一个引号的位置
+                    int endPos = v[0].find_last_of("\""); // 查找最后一个引号的位置
+                    if (startPos != std::string::npos && endPos != std::string::npos && startPos < endPos) {
+                        s2=v[0].substr(startPos + 1, endPos - startPos - 1); // 提取引号中的内容
+                    }
+                    else{std::cout<<"Invalid"<<std::endl;continue;}
+                    booksystem.show_keyword(s2);
+                }
+                else {std::cout<<"Invalid"<<std::endl;continue;}
             }
-            else {std::cout<<"Invalid\n";continue;}
+            else {std::cout<<"Invalid"<<std::endl;continue;}
         }
         else if(command=="buy"){
             if(v.size()==2){
-                if(!isValidInt(v[1])) {std::cout<<"Invalid\n";continue;}
+                if(!isValidInt(v[1])) {std::cout<<"Invalid"<<std::endl;continue;}
                 int tmp=std::stoi(v[1]);
                 double result=booksystem.buy(v[0],tmp);
                 if(result!=0.0){
@@ -132,11 +183,11 @@ int main(){
                     finance.Insert(once);
                 }
             }
-            else {std::cout<<"Invalid\n";continue;}
+            else {std::cout<<"Invalid"<<std::endl;continue;}
         }
         else if(command=="select"){
             if(v.size()==1){booksystem.select(usersystem,v[0]);}
-            else {std::cout<<"Invalid\n";continue;}
+            else {std::cout<<"Invalid"<<std::endl;continue;}
         }
         else if(command=="modify"){
             // 不确定：附加参数内容为空则操作失败
@@ -147,56 +198,80 @@ int main(){
             bool flag=true;
             for(int i=0;i<v.size();++i){           
                 int pos=v[i].find("=");
-                if(pos==std::string::npos) {std::cout<<"Invalid\n";flag=false;break;}
-                if(pos==v[i].size()-1) {std::cout<<"Invalid\n";flag=false;break;}
+                if(pos==std::string::npos) {std::cout<<"Invalid"<<std::endl;flag=false;break;}
+                if(pos==v[i].size()-1) {std::cout<<"Invalid"<<std::endl;flag=false;break;}
                 std::string s1=v[i].substr(0,pos);
-                std::string s2=v[i].substr(pos+1);
                 if(s1=="-ISBN"){
+                    std::string s2=v[i].substr(pos+1);
                     if(already_have[0]=="") {already_have[0]=s2;}
                     else{std::cout<<"Invalid\n";flag=false;break;}
                 }
                 else if(s1=="-name"){
+                    //错误：要去掉引号
+                    //错误：引号在string里面的存储还连带着\符号，因为要说明这是一个普通的字符引号
+                    std::string s2;
+                    int startPos = v[i].find("\"");       // 查找第一个引号的位置
+                    int endPos = v[i].find_last_of("\""); // 查找最后一个引号的位置
+                    if (startPos != std::string::npos && endPos != std::string::npos && startPos < endPos) {
+                        s2=v[i].substr(startPos + 1, endPos - startPos - 1); // 提取引号中的内容
+                    }
+                    else{std::cout<<"Invalid"<<std::endl;flag=false;break;}
                     if(already_have[1]=="") {already_have[1]=s2;}
-                    else{std::cout<<"Invalid\n";flag=false;break;}          
+                    else{std::cout<<"Invalid"<<std::endl;flag=false;break;}
                 }
                 else if(s1=="-author"){
+                    std::string s2;
+                    int startPos = v[i].find("\"");       // 查找第一个引号的位置
+                    int endPos = v[i].find_last_of("\""); // 查找最后一个引号的位置
+                    if (startPos != std::string::npos && endPos != std::string::npos && startPos < endPos) {
+                        s2=v[i].substr(startPos + 1, endPos - startPos - 1); // 提取引号中的内容
+                    }
+                    else{std::cout<<"Invalid"<<std::endl;flag=false;break;}
                     if(already_have[2]=="") {already_have[2]=s2;}
-                    else{std::cout<<"Invalid\n";flag=false;break;}
+                    else{std::cout<<"Invalid"<<std::endl;flag=false;break;}
                 }
                 else if(s1=="-keyword"){
+                    std::string s2;
+                    int startPos = v[i].find("\"");       // 查找第一个引号的位置
+                    int endPos = v[i].find_last_of("\""); // 查找最后一个引号的位置
+                    if (startPos != std::string::npos && endPos != std::string::npos && startPos < endPos) {
+                        s2=v[i].substr(startPos + 1, endPos - startPos - 1); // 提取引号中的内容
+                    }
+                    else{std::cout<<"Invalid"<<std::endl;flag=false;break;}
                     if(already_have[3]=="") {already_have[3]=s2;}
-                    else{std::cout<<"Invalid\n";flag=false;break;}
+                    else{std::cout<<"Invalid"<<std::endl;flag=false;break;}
                 }
                 else if(s1=="-price"){
+                    std::string s2=v[i].substr(pos+1);
                     if(isValidFloat(s2)){
-                        if(already_have[1]=="") {already_have[1]=s2;}
-                        else{std::cout<<"Invalid\n";flag=false;break;}
+                        if(already_have[4]=="") {already_have[4]=s2;}
+                        else{std::cout<<"Invalid"<<std::endl;flag=false;break;}
                     }
-                    else{std::cout<<"Invalid\n";flag=false;break;}
+                    else{std::cout<<"Invalid"<<std::endl;flag=false;break;}
                 }
-                else {std::cout<<"Invalid\n";flag=false;break;}
+                else {std::cout<<"Invalid"<<std::endl;flag=false;break;}
             }
             if(!flag){continue;}
             int tmp = usersystem.login_now_select.back();//当前选中的书的索引
-            if(tmp==-1) {std::cout<<"Invalid\n";continue;}
+            if(tmp==-1) {std::cout<<"Invalid"<<std::endl;continue;}
             Book tmp1;
             booksystem.file_book.seekg(tmp);
             booksystem.file_book.read(reinterpret_cast<char*>(&tmp1),sizeof(Book));
             if(already_have[0]!=""){
                 char exchange[33]={'\0'};
                 std::copy(already_have[0].begin(),already_have[0].end(),exchange);
-                if(std::strcmp(tmp1.isbn,exchange)==0){std::cout<<"Invalid\n";continue;}
-                std::copy(already_have[0].begin(),already_have[0].end(),tmp1.isbn);
+                if(std::strcmp(tmp1.isbn,exchange)==0){std::cout<<"Invalid"<<std::endl;continue;}
                 std::string str(tmp1.isbn,std::strlen(tmp1.isbn));
                 Element e(str);
                 e.value=tmp;
                 booksystem.isbn_chain.Delete(e);
+                //错误：顺序不能放反了
+                std::copy(already_have[0].begin(),already_have[0].end(),tmp1.isbn);
                 for(int i=0;i<70;++i) {e.index[i]='\0';}
                 std::copy(already_have[0].begin(),already_have[0].end(),e.index);
                 booksystem.isbn_chain.Insert(e);
             }
             if(already_have[1]!=""){
-                std::copy(already_have[1].begin(),already_have[1].end(),tmp1.BookName);
                 std::string str(tmp1.BookName,std::strlen(tmp1.BookName));
                 std::vector<int> v1;
                 v1=booksystem.name_chain.Find(str);
@@ -208,9 +283,9 @@ int main(){
                 for(int i=0;i<70;++i) {e.index[i]='\0';}
                 std::copy(already_have[1].begin(),already_have[1].end(),e.index);
                 booksystem.name_chain.Insert(e);
+                std::copy(already_have[1].begin(),already_have[1].end(),tmp1.BookName);
             }
             if(already_have[2]!=""){
-                std::copy(already_have[2].begin(),already_have[2].end(),tmp1.Author);
                 std::string str(tmp1.Author,std::strlen(tmp1.Author));
                 std::vector<int> v1;
                 v1=booksystem.author_chain.Find(str);
@@ -222,29 +297,43 @@ int main(){
                 for(int i=0;i<70;++i) {e.index[i]='\0';}
                 std::copy(already_have[2].begin(),already_have[2].end(),e.index);
                 booksystem.author_chain.Insert(e);
+                std::copy(already_have[2].begin(),already_have[2].end(),tmp1.Author);
             }
             if(already_have[3]!=""){
                 //不确定：keyword没实现(已写)
-                std::copy(already_have[3].begin(),already_have[3].end(),tmp1.Keyword);
-                std::string str(tmp1.Keyword,std::strlen(tmp1.Keyword));
+                //错误：改之后是 v1，是由 already_have[3]得到的
+                //错误：改之前是 v3，是从 tmp1.Keyword里面读出来的
                 std::vector<std::string> v1;
-                v1=booksystem.get_keyword(str);
-                for(int i=0;i<v1.size();++i){
+                v1=get_keyword(already_have[3]);
+                std::string str(tmp1.Keyword,std::strlen(tmp1.Keyword));
+                std::vector<std::string> v3;
+                v3=get_keyword(str);
+                //先删除元素
+                for(int i=0;i<v3.size();++i){
                     std::vector<int> v2;
-                    v2=booksystem.keyword_chain.Find(v1[i]);
-                    Element e(v1[i]);
+                    v2=booksystem.keyword_chain.Find(v3[i]);
+                    Element e(v3[i]);
                     e.value=tmp;
-                    if(!v2.empty()){//如果有，先删除
+                    if(!v2.empty()) {//如果有，先删除
                         booksystem.keyword_chain.Delete(e);
                     }
-                    for(int i=0;i<70;++i) {e.index[i]='\0';}
-                    std::copy(already_have[3].begin(),already_have[3].end(),e.index);
+                }
+                //现在插入元素
+                for(int i=0;i<v1.size();++i){
+                    Element e;
+                    e.value=tmp;
+                    std::copy(v1[i].begin(),v1[i].end(),e.index);
                     booksystem.keyword_chain.Insert(e);
                 }
                 std::copy(already_have[3].begin(),already_have[3].end(),tmp1.Keyword);
-
             }
-            if(already_have[4]!=""){double num=std::stod(already_have[4]);tmp1.price=num;}
+            if(already_have[4]!=""){
+                //错误：保留两位小数
+                double num=std::stod(already_have[4]);
+                std::stringstream ss;
+                ss << std::fixed << std::setprecision(2) << num;
+                ss >> tmp1.price;
+            }
             booksystem.file_book.seekp(tmp);
             booksystem.file_book.write(reinterpret_cast<char*>(&tmp1),sizeof(Book));
         }
@@ -259,7 +348,7 @@ int main(){
                     finance.Insert(once);
                 }
             }
-            else {std::cout<<"Invalid\n";continue;}
+            else {std::cout<<"Invalid"<<std::endl;continue;}
         }
         else if(command=="log"){
             
@@ -269,10 +358,10 @@ int main(){
         }
         else if(command=="quit" || command=="exit"){
             if(v.size()==0){exit(0);}  //不确定
-            else {std::cout<<"Invalid\n";continue;}
+            else {std::cout<<"Invalid"<<std::endl;continue;}
         }
         else{
-            std::cout<<"Invalid\n";continue;
+            std::cout<<"Invalid"<<std::endl;continue;
         }
     }
     return 0;
